requirements.txt
streamlit
firebase-admin
fpdf2
requests